pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-test-assistant"
version = "0.1.0"
description = "Personal execution-oriented AI testing assistant foundation."
requires-python = ">=3.11,<4.0"
dependencies = [
  "langgraph>=0.4,<1",
  "mcp>=1.26,<2",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
test = [
  "pytest>=9,<10",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
